feat: add service generate url with photo

+++ src/services/delete-post-service.ts
@@ -1,6 +1,7 @@
 import { JuniorRepository } from "@/repositories/junior-repository";
 
 interface DeletePostServiceRequest {
+  profileGithub: string;
   id: number;
 }
 
@@ -9,7 +10,13 @@ interface DeletePostServiceResponse {}
 export class DeletePostService {
   constructor(private juniorRepository: JuniorRepository) {}
 
-  async execute({ id }: DeletePostServiceRequest): Promise<any> {
+  async execute({ profileGithub, id }: DeletePostServiceRequest): Promise<any> {
+    const post = await this.juniorRepository.findOne(profileGithub, id);
+
+    if (!post) {
+      throw new Error();
+    }
+
     const response = await this.juniorRepository.delete(id);
     return response;
   }
